+++ CHANGELOG.md
@@ -6,6 +6,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.9] - 2020-08-16
+### Added
+- Restoring backup updates node model on the frontend
+- Making backup updates node model on the frontend
+- User logout
+- Spinner to operation alert
+- Node provision tooltip on node list page
+- Enable SSH after restore (/boot/ssh file method)
+
+### Fixed
+- Restore mondal not showing for nodeless/uploaded backups.
+- Uploaded image permissions
+
 ## [0.1.8] - 2020-08-14
 ### Added
 - CHANGELOG.md
@@ -23,6 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated npm packages.
 
 
-[Unreleased]: https://github.com/knobik/rpi-cluster-pxe/compare/0.1.8...HEAD
+[Unreleased]: https://github.com/knobik/rpi-cluster-pxe/compare/0.1.9...HEAD
+[0.1.9]: https://github.com/knobik/rpi-cluster-pxe/compare/0.1.8...0.1.9
 [0.1.8]: https://github.com/knobik/rpi-cluster-pxe/compare/0.1.7...0.1.8
 [0.1.7]: https://github.com/knobik/rpi-cluster-pxe/compare/0.1.6...0.1.7